+++ aspire_data/__init__.py
@@ -45,7 +45,7 @@ CLI
 """
 from __future__ import annotations
 
-__version__ = "0.8.3"
+__version__ = "0.9.0"
 
 __all__ = [
     "__version__",